Data wrangling from raw data
I use the pandas package to read the tsv document and upload all contents in this document. I use pd. concat function to reshape this table and set a new index. I use langid to ﬁlter non-English contents. I need to use packages like the "re" package to compile the sample regular expression "\w+(?:[-']\w+)?" and ﬁnd words that match this regular expression. I use the Porter stemming algorithm in the natural language processing method to extract the stem of words. Then use the FreqDist method to count the number of occurrences of words. To output bigram words, I use the N-grams package to pair word and word. I use the CountVectorizer library of sklearn in the natural language processing method for word segmentation and encode the words in the text and convert them into a sparse matrix.